requirements.txt
streamlit==1.29.0
streamlit-javascript==0.1.5
google-generativeai==0.3.1
python-dotenv==1.0.0
edge-tts==6.1.9
aiofiles==23.2.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
